+++ CHANGES.rst
@@ -135,6 +135,7 @@ Development
 - Add basic test suite
 - Adjust machinery to run under pytest
 - Add tests for WiFi
+- Add MQTT telemetry test. Needs running MQTT broker.
 
 
 2019-08-19 0.6.0